requirements.txt
fastapi==0.115.8
pydantic==1.10.19
pycountry==24.6.1
uvicorn==0.20.0
jinja2==3.1.3
supabase==0.7.1
starlette==0.45.3
python-multipart==0.0.20
python-dotenv==1.0.1
flask==3.1.0
flask_wtf==1.2.2
gunicorn==23.0.0
qrcode==8.2
reportlab==4.4.2
pillow==11.2.1
cloudinary==1.44.1
